Added useHistory hook
